minor: test and minor edits

+++ ampscz_asana/lib/qc.py
@@ -646,11 +646,14 @@ def get_run_sheet_df(phoenix_dir: Path,
     return datatype_df
 
 
-def dataflow_dpdash(datatype_df: pd.DataFrame, outdir: Path) -> None:
+def dataflow_dpdash(datatype_df: pd.DataFrame,
+                    outdir: Path,
+                    test: bool = False) -> None:
     '''Convert datatype_df to DPDash importable format and save as csv files'''
     # flush existing files
-    for i in outdir.glob('*mridataflow-day*csv'):
-        os.remove(i)
+    if not test:
+        for i in outdir.glob('*mridataflow-day*csv'):
+            os.remove(i)
 
     # loop through each subject to build database
     all_df = pd.DataFrame()
